requirements.txt
fastapi
uvicorn
python-multipart
numpy
scipy
opencv-python
pillow
openai
transformers
torch
timm
scikit-learn
scipy
python-dotenv
pydantic
fastdtw
tqdm
replicate
requests
diffusers
accelerate
sentencepiece
protobuf
imageio-ffmpeg
torchvision
jinja2